+++ crates/profile/models/supporting-types/src/unsecurified_entity.rs
@@ -57,7 +57,7 @@ where
                 unsecured_entity_control,
             )),
             EntitySecurityState::Securified { .. } => {
-                Err(CommonError::AESDecryptionFailed)
+                Err(CommonError::SecurityStateSecurifiedButExpectedUnsecurified)
             }
         }
     }
