+++ crates/system/os/factors/src/apply_security_shield/sargon_os_apply_security_shield_interaction.rs
@@ -28,12 +28,9 @@ impl OsApplySecurityShieldInteraction for SargonOS {
        .unsecurified_erased()
             .iter()
             .map(|e| {
-
                 let provisional = e.entity.get_provisional().expect("Entity should have a provisional config set since we applied shield above");
                 let derived = provisional.as_factor_instances_derived().expect("Should have derived factors");
-
-                let input =                 TransactionManifestApplySecurityShieldUnsecurifiedInput::new(derived.clone());
-
+                let input = TransactionManifestApplySecurityShieldUnsecurifiedInput::new(derived.clone());
                 TransactionManifest::apply_security_shield_for_unsecurified_entity(
                     e,
                     input,
