+++ crates/transaction/manifests/src/lib.rs
@@ -4,6 +4,7 @@ mod delete_account;
 mod high_level;
 mod manifest_account_locker;
 mod manifest_assets_transfers;
+mod manifest_builder_from_manifest;
 mod manifests;
 mod manifests_create_tokens;
 mod manifests_security_shield;
@@ -18,6 +19,7 @@ pub mod prelude {
     pub use crate::high_level::*;
     pub use crate::manifest_account_locker::*;
     pub use crate::manifest_assets_transfers::*;
+    pub use crate::manifest_builder_from_manifest::*;
     pub use crate::manifests::*;
     pub use crate::manifests_create_tokens::*;
     pub use crate::manifests_security_shield::*;
