+++ crates/transaction/manifests/src/manifests_security_shield/mod.rs
@@ -1,0 +1,6 @@
+#[allow(clippy::module_inception)]
+mod manifests_security_shield;
+mod top_up_access_controller_xrd_vault;
+
+pub use manifests_security_shield::*;
+pub use top_up_access_controller_xrd_vault::*;
